First draft for the rti runtime stats module

+++ obc/modules/runtimeTick_mgr/runtimeTick_mgr.h
@@ -5,4 +5,4 @@
 #include <obc_errors.h>
 
 void vConfigureRuntimeStats(void);
-uint64_t vSystemTickGet(void);
+uint32_t vSystemTickGet(void);
